pyproject.toml
[project]
name = "cloudhaven-agent"
version = "0.1.0"
description = "Cloudhaven agent service"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.123.2",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.0.0",
]

